requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
bidict==0.21.2
boto3==1.18.37
botocore==1.21.37
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
charset-normalizer==2.0.4
constantly==15.1.0
cryptography==3.4.8
daphne==3.0.2
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.6
django-allauth==0.44.0
django-cors-headers==3.8.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
google-api-core==2.0.1
google-api-python-client==2.20.0
google-auth==2.0.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
hiredis==2.0.0
httplib2==0.19.1
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
jmespath==0.10.0
msgpack==1.0.2
netifaces==0.10.6
oauthlib==3.1.1
Pillow==8.3.2
protobuf==3.17.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-engineio==4.2.1
python-socketio==5.4.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
zope.interface==5.4.0
